automation/requirements.txt
# Dual Claude Code Orchestration System - Dependencies
# Justice Companion Automation Framework

# Core Dependencies
anthropic>=0.18.0          # Claude API client
watchdog>=3.0.0            # File system monitoring
python-dotenv>=1.0.0       # Environment variable management

# Web Dashboard
flask>=3.0.0
flask-socketio>=5.3.5
eventlet>=0.33.3

# Terminal UI
rich>=13.7.0
textual>=0.47.0

# State Management
filelock>=3.13.0           # Cross-platform file locking (Windows compatible)
redis>=5.0.1               # Optional: distributed state
jsonschema>=4.20.0         # State validation

# Utilities
requests>=2.31.0           # HTTP requests (for webhooks)
pytest>=7.4.3              # Testing framework
